Refactor spec agent to spec skill and update related tests
- Renamed references from "agent" to "skill" in spec generation and routing logic.
- Updated test cases to reflect the new skill structure, including changes in mock implementations.
- Removed deprecated spec agent boot tests and replaced them with skill-specific tests.
- Enhanced spec prompt builders to include updated skill context and instructions.
- Added comprehensive tests for the new spec skill functionality, ensuring proper prompt building and result parsing.
