+++ src/spec-generator.ts
@@ -8,13 +8,13 @@
  *   1. Resolve the datasource (explicit or auto-detected)
  *   2. Fetch issue/file details via the datasource
  *   3. Boot the AI provider
- *   4. For each item, tell the AI agent the target filepath and prompt it
+ *   4. For each item, tell the AI the target filepath and prompt it
  *      to explore the codebase and write the spec directly to disk
  *   5. Verify the spec file was written
  *   6. Push spec content back to the datasource via update()
  *
  * The generated specs stay high-level (WHAT, WHY, HOW) because the
- * planner agent in the dispatch pipeline handles detailed, line-level
+ * planner in the dispatch pipeline handles detailed, line-level
  * implementation planning for each individual task.
  */
 
@@ -30,7 +30,7 @@ export const MB_PER_CONCURRENT_TASK = 500;
 /** Default spec-generation timeout in minutes when not specified by the user. */
 export const DEFAULT_SPEC_TIMEOUT_MIN = 10;
 
-/** Default spec-generation warn-phase duration in minutes (agent receives a "wrap up" message). */
+/** Default spec-generation warn-phase duration in minutes (AI receives a "wrap up" message). */
 export const DEFAULT_SPEC_WARN_MIN = 10;
 
 /** Default spec-generation kill-phase duration in minutes (hard termination after warn). */
@@ -87,7 +87,7 @@ export interface SpecOptions {
   retries?: number;
   /** Spec generation timeout in minutes (default: 10) */
   specTimeout?: number;
-  /** Warn-phase timeout in minutes — agent receives a "wrap up" message after this duration (default: 10) */
+  /** Warn-phase timeout in minutes — AI receives a "wrap up" message after this duration (default: 10) */
   specWarnTimeout?: number;
   /** Kill-phase timeout in minutes — hard termination after the warn phase expires (default: 10) */
   specKillTimeout?: number;
@@ -98,7 +98,7 @@ export interface SpecOptions {
 /**
  * Returns a safe default concurrency: min(cpuCount, freeMB/500), at least 1.
  *
- * Each concurrent agent process (provider runtime) is estimated to consume
+ * Each concurrent provider process (provider runtime) is estimated to consume
  * roughly 500 MB of resident memory. The formula caps parallelism at the
  * lesser of the available CPU count and the number of 500 MB slots that fit
  * in current free memory, ensuring the host is not over-committed. The
@@ -157,7 +157,7 @@ export function isGlobOrFilePath(input: string | string[]): boolean {
 }
 
 /**
- * Post-process raw spec file content written by the AI agent.
+ * Post-process raw spec file content written by the AI.
  *
  * Strips code-fence wrapping, preamble text before the first H1 heading,
  * and postamble text after the last recognized spec section.  Returns the
