+++ src/tests/file-logger-integration.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect, vi, beforeEach, afterEach } from "vitest";
 import type { Task, TaskFile } from "../parser.js";
-import type { PlannerAgent } from "../agents/planner.js";
-import type { AgentResult, PlannerData } from "../agents/types.js";
+import type { Skill } from "../skills/interface.js";
+import type { SkillResult, PlannerData } from "../skills/types.js";
 import type { ProviderInstance } from "../providers/interface.js";
 import type { Datasource, IssueDetails } from "../datasources/interface.js";
 
@@ -22,15 +22,16 @@ const { mocks, TASK_FIXTURE, TASK_FILE_FIXTURE } = vi.hoisted(() => {
     content: "# Test\n\n- [ ] Implement the feature",
   };
 
-  const mockPlan = vi.fn<PlannerAgent["plan"]>();
+  const mockPlan = vi.fn();
   const mockExecute = vi.fn();
   const mockGenerate = vi.fn();
+  const mockDispatch = vi.fn();
   const mockCreateSession = vi.fn<ProviderInstance["createSession"]>().mockResolvedValue("sess-1");
   const mockPrompt = vi.fn<ProviderInstance["prompt"]>().mockResolvedValue("done");
   const mockCleanup = vi.fn().mockResolvedValue(undefined);
 
   return {
-    mocks: { mockPlan, mockExecute, mockGenerate, mockCreateSession, mockPrompt, mockCleanup },
+    mocks: { mockPlan, mockExecute, mockGenerate, mockDispatch, mockCreateSession, mockPrompt, mockCleanup },
     TASK_FIXTURE,
     TASK_FILE_FIXTURE,
   };
@@ -54,36 +55,28 @@ vi.mock("../providers/index.js", () => ({
 }));
 
 vi.mock("../providers/router.js", () => ({
-  routeAllAgents: vi.fn().mockReturnValue({
+  routeAllSkills: vi.fn().mockReturnValue({
     planner: [{ provider: "opencode", model: "claude-haiku-4", priority: 0 }],
     executor: [{ provider: "opencode", model: "claude-sonnet-4-5", priority: 0 }],
     commit: [{ provider: "opencode", model: "claude-haiku-4", priority: 0 }],
   }),
-  routeAgent: vi.fn().mockReturnValue([{ provider: "opencode", model: "claude-sonnet-4-5", priority: 0 }]),
+  routeSkill: vi.fn().mockReturnValue([{ provider: "opencode", model: "claude-sonnet-4-5", priority: 0 }]),
 }));
 
-vi.mock("../agents/planner.js", () => ({
-  boot: vi.fn().mockResolvedValue({
-    name: "planner",
-    plan: mocks.mockPlan,
-    cleanup: vi.fn().mockResolvedValue(undefined),
-  } satisfies PlannerAgent),
+vi.mock("../skills/planner.js", () => ({
+  plannerSkill: { name: "planner", buildPrompt: vi.fn(), parseResult: vi.fn() },
 }));
 
-vi.mock("../agents/executor.js", () => ({
-  boot: vi.fn().mockResolvedValue({
-    name: "executor",
-    execute: mocks.mockExecute,
-    cleanup: vi.fn().mockResolvedValue(undefined),
-  }),
+vi.mock("../skills/executor.js", () => ({
+  executorSkill: { name: "executor", buildPrompt: vi.fn(), parseResult: vi.fn() },
 }));
 
-vi.mock("../agents/commit.js", () => ({
-  boot: vi.fn().mockResolvedValue({
-    name: "commit",
-    generate: mocks.mockGenerate,
-    cleanup: vi.fn().mockResolvedValue(undefined),
-  }),
+vi.mock("../skills/commit.js", () => ({
+  commitSkill: { name: "commit", buildPrompt: vi.fn(), parseResult: vi.fn() },
+}));
+
+vi.mock("../dispatcher.js", () => ({
+  dispatch: mocks.mockDispatch,
 }));
 
 vi.mock("../datasources/index.js", () => ({
@@ -134,6 +127,7 @@ vi.mock("../parser.js", () => ({
   parseTaskFile: vi.fn().mockResolvedValue(TASK_FILE_FIXTURE),
   buildTaskContext: vi.fn().mockReturnValue("filtered context"),
   groupTasksByMode: vi.fn().mockImplementation((tasks: Task[]) => [tasks]),
+  markTaskComplete: vi.fn().mockResolvedValue(undefined),
 }));
 
 vi.mock("../orchestrator/datasource-helpers.js", () => ({
@@ -228,6 +222,14 @@ describe("integration: verbose file logging", () => {
     consoleSpy = vi.spyOn(console, "log").mockImplementation(() => {});
     consoleErrorSpy = vi.spyOn(console, "error").mockImplementation(() => {});
 
+    // Set up dispatch mock routing
+    mocks.mockDispatch.mockImplementation(async (skill: Skill<any, any>, input: any) => {
+      if (skill.name === "planner") return mocks.mockPlan(input);
+      if (skill.name === "executor") return mocks.mockExecute(input);
+      if (skill.name === "commit") return mocks.mockGenerate(input);
+      throw new Error(`Unknown skill: ${skill.name}`);
+    });
+
     mocks.mockPlan.mockResolvedValue({
       data: { prompt: "Execute step 1" },
       success: true,
@@ -239,11 +241,10 @@ describe("integration: verbose file logging", () => {
       durationMs: 100,
     });
     mocks.mockGenerate.mockResolvedValue({
-      commitMessage: "",
-      prTitle: "",
-      prDescription: "",
+      data: null,
       success: false,
       error: "mock: not configured",
+      durationMs: 0,
     });
   });
