+++ src/tests/integration/dispatch-flow.test.ts
@@ -4,10 +4,10 @@ import { join } from "node:path";
 import { tmpdir } from "node:os";
 import { execFileSync } from "node:child_process";
 import type { ProviderInstance } from "../../providers/interface.js";
-import type { PlannerAgent } from "../../agents/planner.js";
-import type { AgentResult, PlannerData, ExecutorData } from "../../agents/types.js";
+import type { Skill } from "../../skills/interface.js";
+import type { SkillResult, PlannerData, ExecutorData } from "../../skills/types.js";
 import type { Task } from "../../parser.js";
-import { markTaskComplete } from "../../parser.js";
+// markTaskComplete is now called by the pipeline itself after dispatch returns success
 
 // ─── Hoisted mock references ────────────────────────────────────────
 
@@ -20,11 +20,10 @@ const { mocks } = vi.hoisted(() => {
     data: { prompt: "Execute the task as described." },
     success: true,
     durationMs: 100,
-  } satisfies AgentResult<PlannerData>);
-  const mockPlannerCleanup = vi.fn().mockResolvedValue(undefined);
+  } satisfies SkillResult<PlannerData>);
 
   const mockExecute = vi.fn();
-  const mockExecutorCleanup = vi.fn().mockResolvedValue(undefined);
+  const mockDispatch = vi.fn();
 
   return {
     mocks: {
@@ -32,9 +31,8 @@ const { mocks } = vi.hoisted(() => {
       mockPrompt,
       mockProviderCleanup,
       mockPlan,
-      mockPlannerCleanup,
       mockExecute,
-      mockExecutorCleanup,
+      mockDispatch,
     },
   };
 });
@@ -56,28 +54,24 @@ vi.mock("../../providers/index.js", () => ({
 }));
 
 vi.mock("../../providers/router.js", () => ({
-  routeAllAgents: vi.fn().mockReturnValue({
+  routeAllSkills: vi.fn().mockReturnValue({
     planner: [{ provider: "opencode", model: "claude-haiku-4", priority: 0 }],
     executor: [{ provider: "opencode", model: "claude-sonnet-4-5", priority: 0 }],
     commit: [{ provider: "opencode", model: "claude-haiku-4", priority: 0 }],
   }),
-  routeAgent: vi.fn().mockReturnValue([{ provider: "opencode", model: "claude-sonnet-4-5", priority: 0 }]),
+  routeSkill: vi.fn().mockReturnValue([{ provider: "opencode", model: "claude-sonnet-4-5", priority: 0 }]),
 }));
 
-vi.mock("../../agents/planner.js", () => ({
-  boot: vi.fn().mockResolvedValue({
-    name: "planner",
-    plan: mocks.mockPlan,
-    cleanup: mocks.mockPlannerCleanup,
-  } satisfies PlannerAgent),
+vi.mock("../../skills/planner.js", () => ({
+  plannerSkill: { name: "planner", buildPrompt: vi.fn(), parseResult: vi.fn() },
 }));
 
-vi.mock("../../agents/executor.js", () => ({
-  boot: vi.fn().mockResolvedValue({
-    name: "executor",
-    execute: mocks.mockExecute,
-    cleanup: mocks.mockExecutorCleanup,
-  }),
+vi.mock("../../skills/executor.js", () => ({
+  executorSkill: { name: "executor", buildPrompt: vi.fn(), parseResult: vi.fn() },
+}));
+
+vi.mock("../../dispatcher.js", () => ({
+  dispatch: mocks.mockDispatch,
 }));
 
 vi.mock("../../tui.js", () => ({
@@ -142,12 +136,18 @@ describe("integration: dispatch pipeline with md datasource", () => {
   beforeEach(() => {
     vi.clearAllMocks();
 
+    // Set up dispatch mock routing
+    mocks.mockDispatch.mockImplementation(async (skill: Skill<any, any>, input: any) => {
+      if (skill.name === "planner") return mocks.mockPlan(input);
+      if (skill.name === "executor") return mocks.mockExecute(input);
+      throw new Error(`Unknown skill: ${skill.name}`);
+    });
+
     // Configure the mock executor to simulate success.
-    // It must call markTaskComplete to update the file (checking off the task),
-    // mimicking what the real executor does.
+    // In the new stateless architecture, the pipeline calls markTaskComplete
+    // after dispatch returns success, so the mock should NOT call it.
     mocks.mockExecute.mockImplementation(
-      async (input: { task: Task; cwd: string; plan: string | null }): Promise<AgentResult<ExecutorData>> => {
-        await markTaskComplete(input.task);
+      async (input: { task: Task; cwd: string; plan: string | null }): Promise<SkillResult<ExecutorData>> => {
         return {
           data: { dispatchResult: { task: input.task, success: true } },
           success: true,
@@ -333,7 +333,7 @@ describe("integration: dispatch pipeline with md datasource", () => {
     Object.defineProperty(process.stdout, "isTTY", { value: true, configurable: true });
 
     let executeCalls = 0;
-    mocks.mockExecute.mockImplementation(async (input: { task: Task; cwd: string; plan: string | null }): Promise<AgentResult<ExecutorData>> => {
+    mocks.mockExecute.mockImplementation(async (input: { task: Task; cwd: string; plan: string | null }): Promise<SkillResult<ExecutorData>> => {
       executeCalls++;
       if (executeCalls <= 4) {
         return {
@@ -343,7 +343,6 @@ describe("integration: dispatch pipeline with md datasource", () => {
           durationMs: 50,
         };
       }
-      await markTaskComplete(input.task);
       return {
         data: { dispatchResult: { task: input.task, success: true } },
         success: true,
